chore: tidy utils
