+++ src/utils/homeCards.ts
@@ -1,0 +1,97 @@
+
+// used in landing page /src/pages/index.tsx
+export const API_CARDS = [
+    {
+      title: "Ultra API",
+      links: [
+        { text: "Docs", href: "/docs/ultra-api/" },
+        { text: "Schemas", href: "/docs/api/ultra-api/" },
+      ],
+    },
+    {
+      title: "Swap API",
+      links: [
+        { text: "Docs", href: "/docs/swap-api/get-quote" },
+        { text: "Schemas", href: "/docs/api/swap-api/" },
+      ],
+    },
+    {
+      title: "Trigger API",
+      links: [
+        { text: "Docs", href: "/docs/trigger-api/create-order" },
+        { text: "Schemas", href: "/docs/api/trigger-api/" },
+      ],
+    },
+    {
+      title: "Recurring API",
+      links: [
+        { text: "Docs", href: "/docs/recurring-api/create-order" },
+        { text: "Schemas", href: "/docs/api/recurring-api/" },
+      ],
+    },
+    {
+      title: "Token API",
+      links: [
+        { text: "Docs", href: "/docs/token-api/" },
+        { text: "Schemas", href: "/docs/api/token-api/" },
+      ],
+    },
+    {
+      title: "Price API",
+      links: [
+        { text: "Docs", href: "/docs/price-api/" },
+        { text: "Schemas", href: "/docs/api/price-api/" },
+      ],
+    },
+  ];
+  
+  export const TOOL_KIT_CARDS = [
+    {
+      title: "Swap Terminal",
+      links: [
+        { text: "Playground", href: "https://terminal.jup.ag/" },
+        { text: "Docs", href: "/docs/tool-kits/swap-terminal" },
+      ],
+    },
+    {
+      title: "Unified Wallet Kit",
+      links: [
+        { text: "Playground", href: "https://unified.jup.ag/" },
+        { text: "Docs", href: "/docs/tool-kits/unified-wallet-kit" },
+      ],
+    },
+    {
+      title: "NPM Package",
+      links: [
+        { text: "NPM", href: "https://www.npmjs.com/package/@jup-ag/api" },
+        {
+          text: "Example",
+          href: "https://github.com/jup-ag/jupiter-quote-api-node/tree/main/example",
+        },
+      ],
+    },
+  ];
+  
+  export const DRWG_CARDS = [
+    {
+      title: "Example Scripts",
+      links: [
+        {
+          text: "Typescript",
+          href: "https://github.com/Jupiter-DevRel/typescript-examples",
+        },
+        {
+          text: "Python",
+          href: "https://github.com/Jupiter-DevRel/python-examples",
+        },
+      ],
+    },
+    {
+      title: "Jupiverse Kit",
+      links: [
+        { text: "Playground", href: "https://www.jupiversekit.xyz/" },
+        { text: "GitHub", href: "https://github.com/dannweeeee/jupiverse-kit" },
+      ],
+    },
+  ];
+
