+++ src/utils/navbarConfigs.ts
@@ -1,0 +1,82 @@
+export const navbarConfigs = {
+  "/docs": [
+    { to: "/docs/",
+      label: "APIs",
+      position: "left",
+      items: [
+        { to: "/docs/", label: "Get Started", position: "left" },
+        { to: "/docs/ultra-api/", label: "Ultra API", position: "left" },
+        { to: "/docs/swap-api/", label: "Swap API", position: "left" },
+        { to: "/docs/trigger-api/", label: "Trigger API", position: "left" },
+        { to: "/docs/recurring-api/", label: "Recurring API", position: "left" },
+        { to: "/docs/token-api/", label: "Token API", position: "left" },
+        { to: "/docs/price-api/", label: "Price API", position: "left" },
+        { to: "/docs/perp-api/", label: "Perp API", position: "left" },
+      ],
+    },
+    { to: "/docs/routing/", label: "Routing Integration", position: "left" },
+    { to: "/docs/tool-kits/", label: "Tool Kits", position: "left" },
+    { to: "/docs/api", label: "API Reference", position: "left" },
+    { to: "/docs/misc/", label: "Misc", position: "left" },
+  ],
+  "/search": [
+    { to: "/docs/",
+      label: "APIs",
+      position: "left",
+      items: [
+        { to: "/docs/", label: "Get Started", position: "left" },
+        { to: "/docs/ultra-api/", label: "Ultra API", position: "left" },
+        { to: "/docs/swap-api/", label: "Swap API", position: "left" },
+        { to: "/docs/trigger-api/", label: "Trigger API", position: "left" },
+        { to: "/docs/recurring-api/", label: "Recurring API", position: "left" },
+        { to: "/docs/token-api/", label: "Token API", position: "left" },
+        { to: "/docs/price-api/", label: "Price API", position: "left" },
+        { to: "/docs/perp-api/", label: "Perp API", position: "left" },
+      ],
+    },
+    { to: "/docs/routing/", label: "Routing Integration", position: "left" },
+    { to: "/docs/tool-kits/", label: "Tool Kits", position: "left" },
+    { to: "/docs/api", label: "API Reference", position: "left" },
+    { to: "/docs/misc/", label: "Misc", position: "left" },
+  ],
+  "/": [
+    { to: "/docs/",
+      label: "APIs",
+      position: "left",
+      items: [
+        { to: "/docs/", label: "Get Started", position: "left" },
+        { to: "/docs/ultra-api/", label: "Ultra API", position: "left" },
+        { to: "/docs/swap-api/", label: "Swap API", position: "left" },
+        { to: "/docs/trigger-api/", label: "Trigger API", position: "left" },
+        { to: "/docs/recurring-api/", label: "Recurring API", position: "left" },
+        { to: "/docs/token-api/", label: "Token API", position: "left" },
+        { to: "/docs/price-api/", label: "Price API", position: "left" },
+        { to: "/docs/perp-api/", label: "Perp API", position: "left" },
+      ],
+    },
+    { to: "/docs/routing/", label: "Routing Integration", position: "left" },
+    { to: "/docs/tool-kits/", label: "Tool Kits", position: "left" },
+    { to: "/docs/api", label: "API Reference", position: "left" },
+    { to: "/docs/misc/", label: "Misc", position: "left" },
+  ],
+  "/ecosystem": [
+    { to: "/docs/",
+      label: "APIs",
+      position: "left",
+      items: [
+        { to: "/docs/", label: "Get Started", position: "left" },
+        { to: "/docs/ultra-api/", label: "Ultra API", position: "left" },
+        { to: "/docs/swap-api/", label: "Swap API", position: "left" },
+        { to: "/docs/trigger-api/", label: "Trigger API", position: "left" },
+        { to: "/docs/recurring-api/", label: "Recurring API", position: "left" },
+        { to: "/docs/token-api/", label: "Token API", position: "left" },
+        { to: "/docs/price-api/", label: "Price API", position: "left" },
+        { to: "/docs/perp-api/", label: "Perp API", position: "left" },
+      ],
+    },
+    { to: "/docs/routing/", label: "Routing Integration", position: "left" },
+    { to: "/docs/tool-kits/", label: "Tool Kits", position: "left" },
+    { to: "/docs/api", label: "API Reference", position: "left" },
+    { to: "/docs/misc/", label: "Misc", position: "left" },
+  ],
+};
